switch to project format
Fixed a bug in the day1 solution where
the list of totals could add the "current
sum" for a given elf to the list each time
a new entry was added e.g. an input like,

100
200
300

could result in a list like,

[300, 200, 100].

Now we reorder the list at each "reset" (empty line).